[Frontend] Display all pharmacies
