+++ pharmacy-frontend/src/app/component/navbar/navbar.component.html
@@ -5,6 +5,9 @@
         </a>
         <ul class="navbar-nav mr-auto mt-2 mt-lg-0">
             <li class="nav-item">
+                <a class="nav-link" [routerLink]="'/home'" routerLinkActive="active">Pharmacies</a>
+            </li>
+            <li class="nav-item">
                 <a class="nav-link" [routerLink]="'/medicines'" routerLinkActive="active">Medicines</a>
             </li>
             <!--Patient-->
