+++ pharmacy-frontend/src/app/component/patient/counseling-schedule/counseling-schedule.component.html
@@ -16,7 +16,7 @@
                 <tr *ngFor="let pharmacist of pharmacists; let i = index;">
                     <td>{{pharmacist.user.name}}</td>
                     <td>{{pharmacist.user.surname}}</td>
-                    <td>DISPLAY WHEN ADDED TO MODEL</td><!-- TODO: Dodati kada se doda rate u model -->
+                    <td><ngb-rating [(rate)]="pharmacist.rate" [max]="5" [readonly]="true"></ngb-rating></td>
                     <td><button class="btn btn-primary" (click)='schedule(pharmacist.user.email)'>Schedule</button></td>
                 </tr>
             </tbody>
