+++ pharmacy-frontend/src/app/component/patient/counseling-search/counseling-search.component.html
@@ -43,7 +43,7 @@
                     <td>{{pharmacy.name}}</td>
                     <td>{{pharmacy.address}}</td>
                     <td>{{pharmacy.counselingPrice}}</td>
-                    <td>ADD RATE WHEN IT IS IMPLEMENTED</td>
+                    <td><ngb-rating [(rate)]="pharmacy.rate" [max]="5" [readonly]="true"></ngb-rating></td>
                     <td><button class="btn btn-primary" [routerLink]="['/patient/counselingSchedule', pharmacy.name]" routerLinkActive="active">Show pharmacists</button></td>
                 </tr>
             </tbody>
