+++ pharmacy-frontend/src/styles.css
@@ -9,6 +9,9 @@ body {
     background-size: cover;
 }
 
+html, body { height: 100%; }
+body { margin: 0; font-family: Roboto, "Helvetica Neue", sans-serif; }
+
 input, select{
     width: 100%;
     padding: 12px 20px;
@@ -29,6 +32,15 @@ input, select{
     width: 100%;
   }
 
+  table{
+    background-color: white;
+  }
+
+  ngb-rating {
+    color: #FFC107;
+    font-size: 30px;
+  }
+
   .page{
     width: 360px;
     padding: 0% 0 0;
@@ -39,11 +51,17 @@ input, select{
     z-index: 1;
     background: #FFFFFF;
     max-width: 560px;
-    margin: 0 auto 100px;
+    margin: 0 auto 20px;
     margin-top: 15px;
     padding: 45px;
     text-align: center;
     box-shadow: 0 0 20px 0 rgba(0, 0, 0, 0.2), 0 5px 5px 0 rgba(0, 0, 0, 0.24);
+  }
+
+  .center{
+    position: relative;
+    margin-top: 15px;
+    text-align: center;
   }
 
   .card{
@@ -52,16 +70,17 @@ input, select{
 
   .cardDisplay{
     display: inline-block;
+    margin: 10px;
+    margin-bottom: 0px;
+    vertical-align: top;
   }
 
   .transparent{
     background-color: rgba(255, 255, 255, 0.8);
   }
 
-
-  table{
-    background-color: white;
+  .clickable:hover{
+    cursor: pointer;
   }
 
-html, body { height: 100%; }
-body { margin: 0; font-family: Roboto, "Helvetica Neue", sans-serif; }
+
